\name{print}
\alias{print.seir}
\title{The SEIR model print method}
\description{
 It is the print method for \code{seir} objects.
}
\usage{
\method{print}{seir}(x, digits=2, ...)
}

\arguments{
  \item{x}{Object of class \code{seir}}
  \item{digits}{Number of digits to print}
  \item{...}{Arguments to pass to other methods}
}
